Fix the content shown before the user connects the wallet
Because of the difference in image size and space, using a flex display was causing the images and typography to not be in the same position across the different tabs. Solve this by setting the specific image size and offset for each image.
The invest tab title was the same as the assets tab title. Now the titles are correct